feat(dashboard-service):ajout de microservice pour dashboard admin all done

+++ command-service/src/main/java/com/ensas/commandservice/dtos/CommandDetailsDto.java
@@ -12,7 +12,7 @@ public class CommandDetailsDto {
     private Long commandId;
     private Long equipmentId;
     private int quantity;
-    private int unitPrice;
+    private double unitPrice;
     private String equipmentName;
     private String equipmentImage;
 }
